Puzzles Completely in
No real testing could be done due to the blocks not being coded completely.

A simple code was implemented to allow the player to interact with the "Friends" and to gain the final keys to end the level. No coding has been done for the friends and their movement.

UI for the collection of keys has been introduced. as well as images for the final keys to be collected by interacting with friends.

the assests used were basic due to the artist not having completed them all as yet.